backend/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
python-dotenv==1.0.0
apify-client==1.7.1
stripe==15.1.0
PyJWT==2.8.0
bcrypt==5.0.0
psycopg2-binary==2.9.9
requests==2.32.2
gunicorn==21.2.0
pytz==2024.1
